feat(settings): Add types to the SensitiveDataClient
Adds some concrete types to the data that is inserted through the
`SensitiveDataClient` so that that we can infer which bits are nullable
and how we are required to handle them in the future.

Left the previous functions `setData` and `getData` with deprecation
notices so we can validate this works with a few changes first and
change the rest over in follow-ups.

Co-authored-by: Beth Rennie <beth@brennie.ca>