fix(settings) - Navigate to settings in a password-reset/sync flow
Because Sync is special and not like any other RP, we don't want to take
users to the `reset_password_verified` where it would tell them to
'Continue to Sync'.

Since this is a niche code path and we don't have a good story now how
to handle these flows, lets go straight to settings.

Fixes FXA-11073